extract helpers to init test file
- makes test cases easier to read
